Move AuthCheck logic to utils
